made suggested code changes
